Add dataset insert function
add functions that take datasets as parameter instead of json. These functions can be
used for example from front end.
- Also move the inserter functions for clarity: "json-parser" does not indicate of any db insertions.
