+++ test/hasura/timetables-data-inserter/hsl/hsl-timetables-data-inserter.spec.ts
@@ -2,8 +2,8 @@ import * as config from '@config';
 import { DbConnection, closeDbConnection, createDbConnection } from '@util/db';
 import { queryTable } from '@util/setup';
 import { hslTimetablesDbSchema } from 'hsl/timetablesdb/datasets/schema';
+import { insertDatasetFromJson } from './data-insert';
 import * as testDatasetJson from './example.json';
-import { insertDatasetFromJson } from './json-parser';
 import {
   HslTimetablesDatasetOutput,
   HslVehicleScheduleFrameOutput,
