+++ test/hasura/timetables-data-inserter/hsl/index.ts
@@ -1,3 +1,4 @@
+export * from './data-insert';
 export * from './dataset';
 export * from './models';
 export * from './table-data';
